-updated dependencies -added sentry -added rate limiter -added seo route (for model page only) -added localization routes -added subscription routes (not ready) -added static pages routes -added category filters routes -added dynamic image watermark opacity -added exporter to Craya (upcoming service) -added jwt authentication -refactored to use repositories and services pattern -refactored validators

+++ database/migrations/1652534822101_model_category_filters.ts
@@ -4,14 +4,16 @@ export default class ModelCategoryFilters extends BaseSchema {
   protected tableName = 'model_category_filters'
 
   public async up () {
-    this.schema.createTable(this.tableName, (table) => {
-      table.bigIncrements('id')
-      table.integer('model_id').unsigned().references('models.id').onDelete('CASCADE')
-      table.integer('filter_id').unsigned().references('category_filters.id').onDelete('CASCADE')
-      table.string('value').notNullable()
-      table.timestamp('created_at', { useTz: true })
-      table.timestamp('updated_at', { useTz: true })
-    })
+    if (!await this.schema.hasTable(this.tableName)) {
+      this.schema.createTable(this.tableName, (table) => {
+        table.bigIncrements('id')
+        table.integer('model_id').unsigned().references('models.id').onDelete('CASCADE')
+        table.integer('filter_id').unsigned().references('category_filters.id').onDelete('CASCADE')
+        table.string('value').notNullable()
+        table.timestamp('created_at', { useTz: true })
+        table.timestamp('updated_at', { useTz: true })
+      })
+    }
   }
 
   public async down () {
